Redefine the structure of lexer model.
